requirements.txt
setuptools
pandas
streamlit
google-generativeai
logging
python-docx